src/requirements.txt
boto3
matplotlib
numpy
python-dotenv
